// src/module/entities/model/Quotas.ts
export class Quotas {
  id?: number;
  produtoCod?: number;
  saldo?: number;
  dtInicial?: Date;
  dtFinal?: Date;
  dtAlteracao?: Date;
  hrAlteracao?: Date;
}
export class QuotasFields {
  id?: boolean;
  produtoCod?: boolean;
  saldo?: boolean;
  dtInicial?: boolean;
  dtFinal?: boolean;
  dtAlteracao?: boolean;
  hrAlteracao?: boolean;
}